// src/api/db/tables.ts
export enum TableNames {
	CART_ITEM_TABLE = "cart_has_item",
	CART_TABLE = "cart",
	CATEGORY_TABLE = "category",
	OPENING_HOURS_TABLE = "opening_hours",
	ORDER_ITEM_TABLE = "order_has_item",
	ORDER_TABLE = "order",
	PAYMENT_METHOD_TABLE = "payment_method",
	PRODUCT_TABLE = "product",
	SHOPLIST_ITEM_TABLE = "shoplist_has_item",
	SHOPLIST_TABLE = "shoplist",
	STORE_TABLE = "store",
	USER_TABLE = "user",
}